Remove the shared mime dir when packaging
On Redhat-based distributions, having an extra mime dir in the
XDG_DATA_DIRS confuses the file dialog into not being able to
retrieve any mimetype. Is it a Qt bug? Is it a Redhat bug? I
don't know, but removing the mimedir makes everything work again,
and we don't use our own mimetype definitions anyway, we have
a fallback for them because QMimeDatabase is too stupid to allow
us to add mimetypes programmatically.

We still want to install our own mimetypes, or rather, I suspect
distributions want us to, so we cannot just skip that. But if
you build on redhat-derived distributions, you'll have to manually
remove the mime dir every time after doing a make install.

BUG:361973