BUG:361973 Remove all our custom mime types
This should fix bug 361973 even for people who build Krita on
their own. It must be a bug somewhere in Qt, and a recent
regression, too. Just stopping to use QMimeDataBase doesn't
fix it, since it's some Qt-internal usage that breaks everything
in QFileDialog if there is a mime directory in the XDG_DATA_DIRS.