Initialize color space of the Group Layers correctly
All the group layers in Krita use the color space of the image
and this color space might be initialized with the profile *after*
all the layers are created. So we need to initialize group layers
as well as usual layers, but not with saved profile, but with the
updated profile of the image.

Fixes T2416
BUG:353708