Fix a crash when saving the image when autosave it ready to save
The crash happened because autosave decided to save at
the moment of time when the user has been saving the
image itself. And given that we save the image in a
background thread, autosave could easily do it.

Now SafeSignalLocker blocks not only the image, but also
a special lock that guard KisDocument from entering the
saving code twice.

BUG:362675
Fixes T2430