Fix saving when multiple view of the same document are open
Basically it happens because the subwindows are removed via deleteLater(),
so at first all the windows get closeEvent() and only after that the start
destructing itself and, therefore, unregistering themselves from KisPart.

Now KisPart just checks if the view is still visible to be counted in the
list of views associated with the document.

BUG:361920
Fixes T2432