[FEATURE] Implement Lock Drawing Angle feature
The feature was suggested by the users in the comments to the
Krita Kickstarter 2016 announcement [0].

This option locks the rotation of the brush to its initial drawing
angle.

Go to Brush 'Settings->Rotation->Drawing Angle' and check 'Lock'
checkbox. The name of the checkbox is not very descriptive, but
it is the best thing we can do under the string freeze :)

[0] - https://geektimes.ru/post/275530/#comment_9247098