Fix wrong positioning of children/popup widgets
In Qt 5.5.1 the "change screen" events do not arrive to the widgets
until they are shown on screen and fully initialized. Therefore, if
we initialize geometry of the main window too early, it will be assigned
with a wrong screen and all the child/popup widgets will start
appearing of a different monitor. In Qt 5.6.0 this problem is resolved,
and you can move the window as much as you like even before showing it.

Fixes T2482
BUG:362025