Fix updating of Fade, Distance, Time sensors.
Basically, the curve option widget wasn't updating when changing the
parameters of the distance, fade and time sensors, which were in a
different widget.

This meant that you would not see the new value in action until you switched
the active curve to a different one and then back again, which caused a lot
of confusion with less computer savy users why the brush instructions didn't
work.