Fix flickering in the end of Lod0 recalculation
The problem happened because openGL textures were painted without a
full update. Basically, we shouldn't switch a texture tile from LodN
to Lod0 mode with a partial update, because some part of the tile still
has an old data, causing that weird flickering.

So now we just issue *all* the GUI updates signals in one go, which
basically updates all the tiles and then paints them on screen.

See a comment in KisSuspendProjectionUpdatesStrokeStrategy.

Fixes T2145
BUG:361448