Refactor loading of keyframes
The loading order is now slightly different. Instead of loading all keyframe data 
in one pass before loading the binary data, both are done in the same pass.
The change is needed for implementing keyframing for masks, where the
selection paint devices (and thus the keyframe channels) are not created 
until the binary data loading pass.