BUG:363087 Show the default swap dir in the preferences dialog
The QStandardPaths code to get a swap dir resulted in an empty
path, which got replaced by QDir::tmpPath() in KisSwappedDataStore.

That's to say; if the user didn't configure a swap location, we'd
still use swap, so that isn't the problem. But we didn't show the
location in the preferences dialog.

So, use the same code to get the temp path for swap in both cases.