Fix: create Bspline coefs data consistenly in C-style
Summary:
As spline->coefs data is created on the heap via posix_memalign()
if HAVE_POSIX_MEMALIGN or also HAVE_SSE2 is defined. So in the
alternative code branch it should be created in C style as well,
using malloc (like in the original code) and not new.
Which then will always match the free(spline->coefs); memory release.

The alternative, doing the release either via delete[] or free depending
on the conditions HAVE_POSIX_MEMALIGN or also HAVE_SSE2, would need
more changes to the code, as there are some places in the imported code
where only one variant is done. Going to malloc also keeps the code closer
to the original one.

Test Plan:
tests still pass on a machine where both #ifndef HAVE_POSIX_MEMALIGN and #ifndef
HAVE_SSE2, so where code change affects the build

Reviewers: dkazakov, rempt, #krita:_next

Reviewed By: rempt, #krita:_next

Differential Revision: https://phabricator.kde.org/D1694