Fix "bended lines" stabilizer problem on Windows
The problem is that on Windows the tablet events are coming in bunches,
not uniformly. Therefore any timing-based smoothing system will not work
out of box.

This patch adds a special class KisStabilizedEventsSampler, that makes
the events uniform. It collects a set of events on a 50ms timeframe and
then distributes it uniformly.

The timeframe size should correlate with the maximum size of the delays
created by the events system. On Windows it is 50ms, on Linux 15-20ms.

The timeframe can be configured with "stabilizerSampleSize" config option.

BUG:362445
Ref T2414
CC:kimageshop@kde.org