Allow auto-enabling the multiproperty widget by just clicking on it
If you open the Properties dialog for multiple layers, the controls
for the properties that vary among the layers (e.g. layer name) will
be initially disabled. Before this patch you should have enabled
them manually by clicking on a checkbox nearby, now it is perfectly
enough just to click on the control itself and it will become enabled.

Ref T106
CC:kimageshop@kde.org