BUG:363845 Do not derefence a deleted point
No idea what Inge actually was doing here, it looks like he
stopped in the middle of a refactoring. But since mTarget is
set in the constructor and in the copy constructor, we don't
need any of this in any case.

Maybe we should remove the whole library, who needs
wmf after all..