Remove caching from DabCache
The thing that seems to be slowing down drawing for large auto-brushes (1000px)
is the allocation of memory and not the rendering of the DAB.
renderMirrorMaskSafe is making a copy of the entire brush, thus defeating the
DabCache.

Later, I'll probably add some sort of memory pool to reduce these large
allocations or, if needed, cache the last few Dabs used.