Enable Gamut color and proofing intent.
The image now stored proofing intent, and the monitor intent is the general
rendering intent.

This also allows us to set the gamut warning color, however, for updating
you currently need to switch profiles as well. I'll make a seperate
generateSoftProofingTransform function, to cache the transform better, so it's
easier to update.

Ref 2337