Add adaptation state slider.
This allows you to set whether or not to adapt the white point, that is,
to set the white point to the screen white(full slider) or to convert all whiteson screen to the white point of the profile when using absolute colorimetric
as intent.

This is useful to simulate paper-colour.

Ref T2337