Show the export configuration widget
Now all we need to do is fix all other export filters other than
png to create an export widget, a default configuration setting
and use them internally, too, instead of hard-coded dialogs in
the export method.

Note: we might also need a currentConfiguration() method in
KisImportExportFilter. We'll need to pass the configuration to
KisImportExportManager all the way down to the actual filter.
And when the split-up is really done, we can make it possible
to pass a filter config file on the command-line for the conversion
system.