[numerical boxes] make use of parsespinboxes in the forms of the tools option widgets.
For the moment the KoUnitSpinBoxes have not been replaced.
Spinboxes instancied in the code but not in a form are not changed.

The main problem that have been spot is that the old way of making
the parsespinboxes red (using a stylesheet) cause style issue in the
tools option widget. A new way of making the spinboxes red has been develloped
that used a Palette and the margins of the inner line edit.

Another thing that have been spot is that Qt Creator will simplifiy the
forms file when edited. While doing so it can collapse a QWidget and a
Layout in just a Layout. But it can be problematic if the widget is called
explicitly in the code. To fix those issue one can add a native="true"
agrument to the QWidget tag mannualy.