[numerical boxes] assert that all the functions from the image and layer menu use numerical boxes.
The only exception is the color background selection box. (Qt native)
The singly problem that have been found is in some layer effect panels, the
value of some spinboxes was reset by some QDial and the expression was
transformed in the actual value before it was possible to write the end of it.

The solution is to adapt a little the setValues slots in the parsespinboxes.