Implement data pool for the chunks in openGL canvas updates
This patch implements KisTextureTileInfoPool. A special object that keeps
track of all the chunks allocated during the openGL canvas updates.

This class is also capable of tracking multiple openGL tile sizes
and color space pixel sizes.

It should fix the most of the memory fragmentation problems,
especially on Windows.

CCBUG:363334,364848