Added ability to enter groups with PgUp/PgDown shortcuts
Bug fix https://bugs.kde.org/show_bug.cgi?id=336899

Summary: Added ability to move to a layer group, if the group is not locked. When entering to the group from above, the top layer is activated. When entering to the group from bellow, the bottom layer is activated.

Reviewers: #krita, dkazakov

Reviewed By: #krita, dkazakov

Subscribers: dkazakov, woltherav

Tags: #krita

Differential Revision: https://phabricator.kde.org/D2210