Create a replacement for KoUnitSpinBox that is able to parse math expression.
This class is more or less a copy of the KoUnitSpinBox class, but that inherit KisDoubleParseSpinBox instead of QDoubleSpinBox.

It was impossible to use classes or functions from the ui library in the widget library du to dependance graph.
So the best solution was to recreate the class in the ui lib.

This class should be a base of work for the future conversions of units functions we want to add in the parser.
From the discussion we had it will continue to be able to display the unit as a suffix,
but should also be able to link with a combobox with a list of units. (still alowing to change the current unit on the keyboard).

Tests where performed with the flake plugin.