Fix an annoying bug whith the specific color selector and the parsespinBoxes.
The spinboxes was reset everytime the internal value change, meaning one can't write a formula with many operations or numbers with more than 1 digit.
The fix imply that as long as the spinbox has the focus and the value has not changed textFromValue return the formula instead of the number.

Some changes were done in the color selector widget to avoid that the double spinboxes grow big when the formula get big then doesn't shrink back when the formula is evaluated.
The fix is simply to tell the spinboxes to stay at a prefered size.