Add KoColorSetWidget, and KoColorPatches for current/previous color
Also rearranged the widget a lot. The KoColorSetWidget is automatically filled with the first palette in the list(which is, as always, concept cookie), and the palette is saved... right now on leave and accept events, which might not be the most intuitive, but at the least it is saved and loaded correctly.

Neither widget has the color management from the displayrenderer, as that would require me editing those directly.

I would like an eye on said save/load code.
Auditors: rempt