Make the custom pattern widget take the selection into account.
Now, if you have a selection, it'll use only that area to retreive the pattern from.

I was thinking of whether it needed an extra toggle, but both Scott and Boud thought it was implicit that having a selection means you want to use it.

This is my first time playing with selection code, so a glance would be nice.

Auditors: dkazakov