Uncheck 'embed profile' by default.
It seems we cannot rely on Firefox and ColorD to manage sRGB images as sRGB images with a regular bog-standard sRGB profile. This is shown previously by an experiment with a PNG that has an sRGB chunk, and a png that has a sRGB profile looking exactly as distorted in Firefox. It is a pity that we have to encourage our users not to embed sRGB profile, but as long as something as simple as showing an sRGB image as an sRGB image(which yes, dear commit message reader, is the same as doing exactly nothing with it) is too complicated for other programs, it cannot be helped.

CCBUG:367832