Let the unitParseSpinBox use document relative units (px, vw, vh for the Lenght. s and % for the time).
The main problem that was faced in the process is that the function to reach the corrent view (thus current document and image) are in the ui lib, which is unreachable from widgets.
The solution is to transform the KisSpinBoxUnitManager to some kind of abstract interface. Then subclass it in ui instead of widget.

The only problem then, is that when you create you spinbox (in a place that can probably use kritaui), you need to attach the proper KisSpinBoxUnitManager to it, because it can't be done internally.

The advantage is that, by subclassing KisSpinBoxManager, one can create a lots of different behaviour for different usecases, and the UnitParseSpinBox can more easily be re-used outside of Krita.