Suppress the use of KoStrokeConfig widget in a few krita tool.
This widget is useless when painting on a raster layer, and on a vector layer the stroke can be configured using the default tool.

For karbon tools it's harder. It seems hard to suppress the stroke config, or to set it with a DocumentAwareUnitManager.

There's maybe a solution at higher level, I should check.