Fix "undefined referenced to..." compile errors from KoDualColorButton in Windows
Summary:
KoDualColorButton was initially part of widgets library, but was moved to the ui library in rKRITA93ef7c8

However, it is still marked for shared library export for widgets library instead of ui library. Because of this the compiler cannot find the implementation of KoDualColorButton, and outputs an "undefined reference" error. This patch corrects KoDualColorButton to be exported as part of ui library.

It's odd that the error only shows up in Windows (windows 10 64-bit, TDM-GCC 5.1.0). In linux it compiles just fine (though it probably shouldn't).

Reviewers: rempt, woltherav

Reviewed By: woltherav

Tags: #krita

Differential Revision: https://phabricator.kde.org/D2736