BUG:368648 Do not deadlock while loading this image
The image contains a filter layer that needs create a lab colorspace
for filtering. That locks the KoColorSpaceRegistry, but... We also
need access to load the profiles. And in any case, the image should
not be recomputed while loading.

So, lock the image update while loading. This has the side effect
of noticably speeding up the loading of this image for me. Which
lock to use, Lock, barrierlock, lock or blockupdates... That's a
mystery, because no apidox. I used blockUpdates because that sounds
most like what I need.