Fix the tablet-no-pressure bug when closing the brush editor
1) When closing a popup widget by clicking on some area outside the
application window, we get only a tablet press event, but no tablet
release event, because the release is delivered to some other
application, not us. Therefore we get stuck in an inconsistent state
(we think some buttons are pressed while they are not). I implemented
a method for requesting button state from XI2 event, so now we can sync
with it. The only trouble is that I don't know if those buttons are
logical or physical, so right now the syncing happens *only* when no
button is pressed according to XI2. Otherwise we just give a warning.

2) This patch also fixes QXcbConnection::windowFromId() to lazily
populate window handles map. Usually we do it in Enter event, but
sometimes we don't have this event. For example, when a popup window
opens and the cursor "magically" appears above its surface, we don't
get Enter event. It seems to be quite recent change, because it used
to work before.

Fixes 369114