BUG:369480 Make the toggle-to-previous-preset action work again
It's inevitable to do that check -- we have several preset selector
widgets, and selecting a preset, sets the current preset on all the other
widgets, which then emit a signal to tell the world the preset has been
changed. It would be good if that didn't happen, and it's possible to
achieve that by blocking signals, but that's not robust: it's always
possible something else selects a preset and starts the updating.

Closes T3893