Be more consistent in usage of override keyword
When a member function in a class uses override, all other functions
that override virtuals should follow. This makes refactoring easier
(renaming the base class function for example will result in an error in
the derrived class).
Most compilers have a warning for the use of override in only some
functions. In order to reduce compiler warnings, add missing overrides.

The warnings look like:
warning: 'icon' overrides a member function but is not marked 'override'
[-Winconsistent-missing-override]

This patch does not touch KisLayer::original() since that is actually
weird:
in KisBaseNode original is a virtual function (not pure) while in the
derrived class it's declared as pure.