Fix the Predefined Brushes
The problem is that we share a single KisBrush object across the entire
Krita. Therefore, if we start writing into the settings object and some
other code will get notifications that "something has changed", that other
location may override some settings in the shared KisBrush. Therefore we
will continue to write wrong values to the brush.

Therefore now we postpone all the updates of the settings object until the
entire write operation is completed. The postponing is implemented using a
special KisPaintOpPreset::UpdatedPostponer class that uses the interface
of updates proxy to do the stuff

Fixes T3925
BUG:369532