Implemented Undo for Instant Preview strokes
Now the user can cancel a stroke without waiting until the main
full-resolution stroke finishes its execution. Just press Ctrl+Z
and the your preview will be (partially) reverted and the corresponding
background stroke will be cancelled.

Technically it is implemented by introducing the second undo store
inside the strokes queue and syncing it with the running strokes
with a set of complicated rules. If the stroke can be cancelled,
it is cancelled, if the stroke has already completed its execution,
it is just undone.

BUG:363311
Ref T2544