BUG:371463 Don't crash when creating an image from a template
Sometimes, if the dialog would open by default on a template
page, and you'd create an image from a template, the dialog
and pane would be deleted too early. We said, when creating
the slots, that deleteLater would be used to delete the dialog,
but we didn't do that... Now we do.