Copy KF5 Qt plugins from another possible location
The KDE Framework build files checks whether the Qt install prefix
is the same as the KF5 install prefix and sets the default value of
KDE_INSTALL_USE_QT_SYS_PATHS accordingly, which affects where the plugins
are copied to.  As a result, the file locations differ whether Qt is
built separately or together with other 3rdparty stuff.  This tries to
copy from both locations.