Remove strange list check
My best guess is that the check was meant to be for !children().empty(),
in which case it would try to optimize calling qDeleteAll away.

The check for at(0) would actually crash if the list was ever empty, so
the qDeleteAll must have been executed unconditionally before.
Have less code and less potential for crashing.