cmake: use cmake-extra-modules for Unit Tests
Summary:
This actually makes it less boiler plate heavy.
It also fixes a few small issues with the tests, three tests were
previously forgotten and have been re-added.

With the custom krita_add_broken_unit_test the same functionallty is
preserved that was there before in the fork of the kde4 macro.
This is a step towards reducing the amount of custom cmake code that
Krita drags around and will make the build system easier.

Thanks to the new macros, the total amount of redundant macro code is
reduced and we lose a bit over 1000 lines of useless boilerplate.

The total test count goes up by three.

The flake tests had their data in the source path, which is unlike any
other tests and were relying on some old defines. They have been adapted
to behave like any other test.

Reviewers: #krita

Subscribers: rempt

Differential Revision: https://phabricator.kde.org/D3167