cmake: Clean X11 and QtX11Extras detection
Qt5X11Extras is already required a few lines above.
Xinput would fail the build if it's not found below. That in turn means
we need X11 as well. When getting rid of X11 dependencies for Wayland
and friends it will be easier to start with this simple set of rules.