Fix crashes in krita-paintop-EmbeddedPatternManagerTest
KisSharedPtr proudly points out that it doesn't have an explicit
constructor. Maybe that's not so great though: the implicit conversion
to smart pointers means that suddenly objects on the stack get
implicitly passed to a shared pointer and are definitively double
deleted.

The test still fails, only the double delete has been fixed.