Fail if we cannot lock the image or for saving, but do not show an error dialog
This moves the locker object up in saveFile, and returns false if
we cannot lock. Image locking failures at this stage can only be
because we try to autosave while adding a lot of strokes if lod
is active; we don't want to show error boxes, but silently retry
in 10 seconds.

(The case where the user explicitly saves but the image is busy
is caught in KisMainWindow).