BUG:370566 Save a copy of the current image
This gets rid of both the delayed save dialog in KisMainWindow
(which offered a cancel button that didn't work to cancel the
runnin strokes), and the locking in KisDocument (which dead-locked
Krita and caused dataloss).

Instead, a shallow clone of the image is created that can be safely
used to save. The strokes will run on the original image, which means
that the result of still-running strokes is not saved. The user is
warned about that. (That means a new message has been added.)