Partially revert "Fix bad weak pointer usage"
These changes are impacting the creation of images and show how fragile
the construction is. Inside the image constructor, weak pointers to the
image are created, without a single strong reference. When one of the
weak pointers is converted to a strong ref, it will be the only
reference, once that goes out of scope, the image is deleted, inside the
constructor. This means more thinking has to go into the whole ownership
model, but for now get rid of the crash.

This partially reverts commit ac2868ad23b50b5e3196eb0d5d1302f5bb02c741.