Work around low framerate of stabilizer
Work around low framerate of stabilizer by delaying and painting the
stroke progressively.

The `KisStabilizerDelayedPaintHelper` class collects the sampled events
and distributes them evenly with a timestamp attached, then a timer firing
at a closer interval would paint the line bit by bit in order to give a
smoother user feedback.

The config option `stabilizerDelayedPaintInterval` with default value `20`
controls the paint interval. The delayed painting is disabled if this
value is higher than `stabilizerSampleSize`, which means it is disabled by
default on non-Windows system.

BUG: 369349
CLose: T4182
Differential Revision: https://phabricator.kde.org/D3321