Fix isotropic spacing when working on a mirrored canvas
Summary: BUG:372545

Test Plan:
This patch touches rotation and mirroring capabilities of the brushes. These
parameters depend on both, brush and canvas rotation/mirroring capabilities,
so they should be tested in complex.

1) Test with rotated (or drawing angle) brushes
2) Test with mirrored (including using sensor) brushes
3) Test with rotated canvas
4) Test with mirrored canvas
5) Test with all combinations of the above :)

The general idea of canvas rotation/mirroring is:

Whenever you rotate or mirror the canvas, brush shape *doesn't* change
its shape from your point of view. That is if you paint with a
vertical rectangular brush, the brush should still stay vertical
whenever you rotate/miror the canvas. In other words, the brush shape
is linked to your stylus tip, not to the canvas.

Reviewers: #krita, woltherav, timotheegiet

Differential Revision: https://phabricator.kde.org/D3420