All shortcuts should be configurable when no document is open
The problem happens because some actions are created/added-to-collection
only after a document is created. Now we workaround it by faking the
absent actions and deleting them together with KisActionsSnapshot.

Another problem is that some of the actions, e.g. "show_color_history",
is created long before it is added to the global action collection. It
is _created_ right at the start of Krita, but _added_ only when a new document
is constructed. If we change the shortcut between these two events, then
the changed value will be lost. Now we workaround this problem by connecting
to the action collection, listening to 'inserted()' signal and updating
shortcuts when an action is added to the collection.

BUG:370720