Move mikro out of the way
Mikro was an interesting experiment, and the version we have works
better than the one the Scribus community is experimenting with,
but it's still too fragile. This means we stop providing api in the
form of Q_PROPERTIES, only methods, signals and slots are possible
now.

The central core of the scripting api (Krita.instance() is aliased
twice: as Scripter and as Application. The first is familiar for
people who are used to Scribus scripting api, the second follows
Photoshop's api.