Fix ugly rendering of shape handles and bounding boxes
Finally! :)

Now there is a special class for rendering handles in native
display coordinates, instead of the transformed one. Its name is
KisHandlePainterHelper. It uses KisAlgebra2D::DecomposedMatix
angorithm, which decomposes an arbitrary matrix into a set of
primitive transformations. Then we use shear and rotate parts of it
to transform the handles and drop the rest :)