[FEATURE] Ctrl-key continued mode for Outline Selection Tool
This is a special mode of the outline selection tool. If
you press Ctrl while dragging the cursor, you can raise
the stylus from the tablet surface and have a bit of rest.
Then just continue the selection from the arbitrary point
and the new selection segment will be connected to the
older one with a straight line.

This shortcut basically temporary switches outline tool
into polygonal tool. That is extremely useful for painters
who do flat-filled art.

CC:kimageshop@kde.org