CCBUG:373821 Try to avoid a crash  in the text editor
This seems to happen in the change tracker, according to the
backtrace for this bug. Well, we don't need no changetracker in
Krita (it's not the thing that's used for undo), so remove all
this highly complex code and let's hope that tonight's nightly
build on https://ci.appveyor.com/project/alvinhochun/krita/build/artifacts
doesn't crash anymore.