Implement global scaling mode
In Global Scaling Mode the spin boxes in the geometry box
measure the size of your shape in *absolute* coordinates, that is
in X,Y coordinates of your image. That way you can find out the
bounding box of the heavily transformed shapes.

Global Scaling Mode can also be combined with "non-uniform scaling"
mode. It this case, changing width or height of the shapes will keep the
geometry of the shape(!), that is will not add any shears to the rotated
shapes. If you combine Global Scaling Mode and Uniform Scaling, then
scaling rotated shapes will shear during the transformation.

PS:
I feel these options will need quite a bit of video
documentation/explanation!

CC:kimageshop@kde.org