CCBUG:375253 Replace asserts with warnings.
Despite the asserts being of the KIS_ASSERT_RECOVER_NOOP type,
they lead to a hard assert without a message being shown in our
windows builds. I do not know why that is, but it means that this
check gives a crash without any information.

So, for now, let's disable the asserts so we can test what is really
going on.