Refactor the resize canva dialog to use unitparsespinboxes
Ref T1621

The new dialog looks more or less the same, except now it's possible to change the unit with the keyboard (and new units are present, like vh or vw).

For the moment, only the resize canva dialog have been refactored. If you want to test the new design and propose other approaches test with this dialog.

Mayba a good idea would be to have a dedicated widget with the spinbox and the combobox. Then the user can select via the configuration if he/she want the unit to be displayed in the spinbox (and the combobox is hidden) or in the combobox. What do you think ?