BUG:374928 Fix loading the translations that fit the system locale
It is supremely important NOT to use i18n before the application
is created.  If i18n is called too early, like when creating a static
KoID, the translation system is initialized before we've had a chance
to add the path to the translations to the XDG_DATA_DIRS variable. Then
the translation system only sees the globally available translations,
and Krita stays untranslated.

Always use ki18n when initializing a static object with a translation.

CCMAIL:kimageshop@kde.org