Implement auto-fill function for Krita markers
Now we define a custom SVG attribute "krita:marker-fill-method=auto",
which means that all the markers should have the same fill (including
gradient and patterns) as the shape's stroke.

Ideally we should make a GUI switch, so the user would be able to choose
whether he wants to auto-fill the markers or not. But that can wait till
later.

CC:kimageshop@kde.org