BUG:376993 Don't crash when a file layer is deleted
QFileSystemWatcher cannot be deleted in a different thread from
where it was created. That did happen when a file layer was deleted
after converting it to a paint layer when the undo command was
pushed off the stack.

This change creates a single static QFileSystemWatcher that all
file layers connect to. fileChanged checks that the changed file
is the right one before setting the fileChanged flag.