Fixed a lot of double-rendering problems in Flake
The access to the shapes RTree was not consistent. Some shapes were
**never** officially added to the tree and resulted in being added
using the spontaneous update calls. It resulted them in being added
to the tree multiple times, which called the double rendering with all
the consequences.

This patch also adds two sanity checks to KoRTree to catch such
problems automatically in the future.