Disable copying the image before saving
The copy is incomplete, the copy doesn't copy the projection so
we need to re-render the image, which makes saving very slow, and
we're not saving in the background at this moment, so the copy
isn't useful.

Note: I have intentionally left the code intact, and added a warning.
Saving in the background is still todo.

CCBUG:375585, 374345