Refactor the imagesize and imageoffset plugins to use the new unitparsespinboxes.
Not so much has changed... apparently. But now you're able to change the
unit with the keyboard (even if the old method, based on comboboxes, is
still present... which is usefull for thoses who use the mouse but also
to know the list of possible units).

Some new fancy units have been added, like view width (vw) and view
height (vh) that represent 1% of the width, respectivly the height of
the image.